requirements.txt
selenium>=4.15.0
google-generativeai>=0.3.0
pydantic>=2.0.0
